refactor: Rename ansible.el to ansible-mode.el
This commit refactors the ansible.el file to ansible-mode.el for
clarity and consistency with minor-mode naming conventions in Emacs.
All references within the code were updated to comply with this change,
including variable and function names.

For example:
- `(require 'ansible)` is updated to `(require 'ansible-mode)`
- Vars and faces updated from `ansible-*` to `ansible-mode-*`.

Additionally, `ansible-yasnippet.el` is split out to optionally
handle Ansible snippets.

For now this drops the vault and playbook listing commands.